# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Informasi Manajemen Inventaris Sekolah API",
        "description": "REST API untuk Sistem Informasi Manajemen Inventaris Sekolah. Mendukung modul Auth, Pengguna, Peran & Akses, Master Sekolah, Master Barang, Manajemen Aset, Transaksi Peminjaman & Permintaan, Pemeliharaan & Akhir Siklus Aset, serta Manajemen Database.",
        "contact": {
            "name": "Admin Sekolah",
            "email": "admin@sekolah.sch.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login pengguna",
                "description": "Autentikasi pengguna menggunakan username dan password, mengembalikan token Sanctum.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Username atau password salah"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout pengguna",
                "description": "Revoke token Sanctum yang sedang digunakan.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logout berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout berhasil."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid atau tidak diberikan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Profil pengguna yang sedang login",
                "description": "Mengembalikan data pengguna yang sedang login beserta relasi peran, akses, kelas, mapel, dan unit.",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Data pengguna berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid atau tidak diberikan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AksesResource": {
                "description": "Representasi data modul akses / permission",
                "properties": {
                    "id_akses": {
                        "description": "============================================================\n SCHEMA DEFINITIONS – Akses Module\n============================================================",
                        "type": "integer",
                        "example": 1
                    },
                    "nama_modul": {
                        "type": "string",
                        "example": "inventaris"
                    },
                    "hak_buat": {
                        "type": "boolean",
                        "example": true
                    },
                    "hak_baca": {
                        "type": "boolean",
                        "example": true
                    },
                    "hak_ubah": {
                        "type": "boolean",
                        "example": true
                    },
                    "hak_hapus": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
            "LoginRequest": {
                "description": "Payload untuk login pengguna",
                "required": [
                    "username",
                    "password"
                ],
                "properties": {
                    "username": {
                        "description": "============================================================\n SCHEMA DEFINITIONS – Auth Module\n============================================================",
                        "type": "string",
                        "example": "admin"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "LoginData": {
                "description": "Data hasil login berisi pengguna dan token",
                "properties": {
                    "pengguna": {
                        "type": "object"
                    },
                    "token": {
                        "type": "string",
                        "example": "1|abc123tokenxyz"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "description": "Response wrapper untuk login berhasil",
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Login berhasil."
                    },
                    "data": {
                        "$ref": "#/components/schemas/LoginData"
                    }
                },
                "type": "object"
            },
            "MeResponse": {
                "description": "Response wrapper untuk data pengguna yang sedang login",
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Data pengguna berhasil diambil."
                    },
                    "data": {
                        "description": "Objek data pengguna beserta relasinya",
                        "properties": {
                            "id_pengguna": {
                                "type": "integer",
                                "example": 1
                            },
                            "username": {
                                "type": "string",
                                "example": "admin_rpl"
                            },
                            "id_peran": {
                                "type": "integer",
                                "example": 1,
                                "nullable": true
                            },
                            "id_kelas": {
                                "type": "integer",
                                "example": null,
                                "nullable": true
                            },
                            "id_mapel": {
                                "type": "integer",
                                "example": null,
                                "nullable": true
                            },
                            "id_unit": {
                                "type": "integer",
                                "example": null,
                                "nullable": true
                            },
                            "peran": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/PeranResource"
                                    }
                                ],
                                "nullable": true,
                                "type": "object"
                            },
                            "kelas": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/KelasResource"
                                    }
                                ],
                                "nullable": true,
                                "type": "object"
                            },
                            "mapel": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/MapelResource"
                                    }
                                ],
                                "nullable": true,
                                "type": "object"
                            },
                            "unit": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/UnitResource"
                                    }
                                ],
                                "nullable": true,
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "KelasResource": {
                "description": "Representasi data kelas",
                "properties": {
                    "id_kelas": {
                        "description": "============================================================\n SCHEMA DEFINITIONS – Kelas Module\n============================================================",
                        "type": "integer",
                        "example": 1
                    },
                    "nama_kelas": {
                        "type": "string",
                        "example": "Kelas A"
                    },
                    "id_rombel": {
                        "type": "integer",
                        "example": 1
                    },
                    "rombel": {
                        "properties": {
                            "id_rombel": {
                                "type": "integer",
                                "example": 1
                            },
                            "nama_rombel": {
                                "type": "string",
                                "example": "X TKJ 1"
                            },
                            "id_jurusan": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "MapelResource": {
                "description": "Representasi data mata pelajaran",
                "properties": {
                    "id_mapel": {
                        "description": "============================================================\n SCHEMA DEFINITIONS – Mapel Module\n============================================================",
                        "type": "integer",
                        "example": 1
                    },
                    "nama_mapel": {
                        "type": "string",
                        "example": "Matematika"
                    }
                },
                "type": "object"
            },
...
            "PeranResource": {
                "properties": {
                    "id_peran": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_peran": {
                        "type": "string",
                        "example": "Administrator"
                    },
                    "akses_list": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AksesResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "UnitResource": {
                "properties": {
                    "id_unit": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_unit": {
                        "type": "string",
                        "example": "Unit Sarana Prasarana"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}